tmemes: stop unpacking static files into the data directory
Now that we are reading and parsing the font TTF directly instead of expecting
the library to do it under the covers, we no longer need to unpack files to the
"real" filesystem, but can read/serve them out of the embed.

Update the /static/ handler to serve from there too.