tmemes: remove font cache
It turns out image/font.Face implementations are not required to be safe for
concurrent use by multiple goroutines. The truetype.Face implementation, in
particular, keeps a mutable cache of glyph metrics. When regenerating multiple
images that shared a point size, this could lead the handler to panic.

It is, however, safe to cache the parsed *truetype.Font, which does not have
any mutable state, and is the more expensive step of face construction.

Fixes #1