types: add Areas type and use it for fields
Areas is a slice of Area, but it supports decoding JSON as either a single
object (the current format) or an array of multiple areas (to support the
ability to move text frame by frame).

Also add a hold count so we can do something about cycle timing.  I don't want
to deal with tweening, but we might have to.