types,tmemes: add support for tweening moving text
Add a "tween" field to Area that controls whether the text is stationary (the
default) or moves toward its next location (if true). Tweening is done by
linear interpolation toward the next area in sequence.

The rendering sometimes leaves a trail of artifacts, but I suspect we can
probably find a way to mask those out.