types,tmemes: add support for timed captions
Add start and end fields to TextLine, representing fractions of the frame
length of an animation. If start > 0, the line will not be made visible before
that point in the animation. If end > start, the line will be hidden after that
point in the animation. If start == end == 0, the default, the line will show
for the whole animation.

Also, take this opportunity to split a bunch of the helper code that had
accumulated in ui.go, but was not related to the UI, into its own file.