tmemes/static: attempt at a row-major masonry layout using css only (#2)
This uses some hacks I learned from:

https://tobiasahlin.com/blog/masonry-with-css/

It's not perfect, since we have to define the height of the meme-list
container. I added a max-height for cards and worked out an upper
bound for the height.

It's also limited to 3 columns. Apparently we can do more, but that
relies on inserting tags in the html, which is a step too far.