tmemes/static/style.css: kludge masonry layout it until we do pagination
The masonry layout CSS does some math on the maximum number of items
per page to place an upper bound on the meme container element
height. Since we don't do pagination yet, we can either set that
with javascript, defeating the point of using pure CSS masonry, or
bump its reckoning of "memes per page" to a very large number.

The down side of doing the latter is that we get a bunch of weird
space at the bottom of the page.  More motivation to paginate, I
guess.