style.css: remove masonry hackery and switch back to a grid
It's not as pretty as masonry, but it's less hacky.

We can have it back with some javascript (e.g. [1]) or wait until
native CSS masonry layout is implemented by browsers.

Since this fixes the weird spacing that was a side effect of the
previous approach, this moves the next page/prev page links to the
bottom of the page, where they belong.

[1] https://medium.com/@andybarefoot/a-masonry-style-layout-using-css-grid-8c663d355ebb)
[2] https://developer.mozilla.org/en-US/docs/Web/CSS/CSS_grid_layout/Masonry_layout